broke layout huge component into small components
